% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cross.r
\name{crossFUN}
\alias{crossFUN}
\title{A function to set up data ready to perform "cross" analysis}
\usage{
crossFUN(x, group1, group2, date.start = "2016-07-01", date.end = NA)
}
\arguments{
\item{x}{This is the column under observation.}

\item{group1}{This is the id of each set of data being compared.}

\item{group2}{This is the column which identifies separate networks within the data.}

\item{date.start}{This sets the start of the selected data. Defaults to '2016-07-01'.}

\item{date.end}{This sets the end of the selected data. Defaults to one week later than `date.start`.}
}
\description{
This function reformats the data relative to cross analysis, where time is held steady. NB: id & group1 here are the same, and so only id is included.
}
\examples{
crossFUN()
}
