% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{distFUN}
\alias{distFUN}
\title{A selection of functions used within other functions in the package}
\usage{
distFUN(x, d.thresh = 0.3, plot = TRUE)
}
\arguments{
\item{d.thresh}{This is the dissimilarity threshold (1-p).}

\item{plot}{This decides whether to plot the data. Defaults to TRUE.}
}
\description{
A function for creating dissimilarity plots using Spearman rank correlations.
}
\examples{
distFUN()
}
