% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idcFUN.R
\name{idcFUN}
\alias{idcFUN}
\title{An inverse distance correlation function.}
\usage{
idcFUN(x, obs = "obs", date = "date", group, lat, lon, reflective = TRUE,
  theta = 0.5, tau = 120, window.size = 72)
}
\arguments{
\item{obs}{The column under examination.}

\item{date}{A date column (set up as YYYY-MM-DD HH:MM:SS).}

\item{group}{This is the column where data will be partitioned upon (either site or time).}

\item{lat}{Latitude.}

\item{lon}{Longitdue.}

\item{reflective}{This decides whether to use a selected time period of data (reflective, `TRUE`, the default setting), or to use the data from the latest day (live, `FALSE`).}

\item{theta}{This sets the test threshold on whether to flag the data. Defaults to `NA` (no flags given).}

\item{tau}{This sets the day threshold on whether to flag the data given consistent theta flags. Defaults to `NA` (no flags given).}
}
\description{
This function compares the correlations of one site against the other sites in the network. The final product is a sum of the correlations multiplied by the inverse distance. Data are rolled iteratively to make a continuous examination of network correlations over time.
}
\examples{
idcFUN()
}
