% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{ksplotFUN}
\alias{ksplotFUN}
\title{This function provides the KS plot.}
\usage{
ksplotFUN(x, obs, proxy, site)
}
\arguments{
\item{obs}{This is the column being examined.}

\item{proxy}{This is the comparison variable}

\item{site}{This is the grouping factor}
}
\description{
This function provides the KS plot.
}
\examples{
ksplotFUN()
}
