% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{nearestsiteFUN}
\alias{nearestsiteFUN}
\title{This function gives the proxy type of nearest site data}
\usage{
nearestsiteFUN(x, meta, site, type)
}
\arguments{
\item{site}{This is the grouping variable}

\item{type}{A second grouping variable}

\item{obs}{This is the column being examined.}

\item{plot}{This decides whether to plot the data. Defaults to TRUE.}
}
\description{
This function gives the proxy type of nearest site data
}
\examples{
nearestsiteFUN()
}
