% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{networkmedianFUN}
\alias{networkmedianFUN}
\title{This function gives the proxy type of network median}
\usage{
networkmedianFUN(x, obs, site, type)
}
\arguments{
\item{obs}{This is the column being examined.}

\item{site}{This is the grouping variable}

\item{type}{A second grouping variable}

\item{plot}{This decides whether to plot the data. Defaults to TRUE.}
}
\description{
This function gives the proxy type of network median
}
\examples{
networkmedianFUN()
}
