% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentFUN.r
\name{percentFUN}
\alias{percentFUN}
\title{A function for comparing the daily minima and maxima.}
\usage{
percentFUN(x, date, obs, group, reflective = TRUE,
  date.start = "2016-07-01 00:00:00", date.end = "2016-07-10 00:00:00",
  theta = 2, tau = NA)
}
\arguments{
\item{date}{This is a required column with the format 'YYYY-MM-DD HH:MM:SS'.}

\item{obs}{This is the numeric column under investigation.}

\item{group}{This is the first grouping value of the data, with the default being 'site'.}

\item{reflective}{This decides whether to use a selected time period of data (reflective), or to use data from the latest day (live). Defaults to TRUE.}

\item{date.start}{This sets the start of the selected data, if reflective is true. Defaults to '2016-07-01 00:00:00'.}

\item{date.end}{This sets the end of the selected data, if reflective is true. Defaults to '2016-07-10 00:00:00'.}

\item{theta}{This sets the test threshold on whether to flag the data. Defaults to 2 (similar to 2 standard deviations).}

\item{tau}{This sets the day threshold on whether to flag the data given consistent theta flags. Defaults to NA (no flags given).
export
examples
percentFUN()}
}
\description{
This function finds and compares the daily 'minima' (5th) and 'maxima' (95th) percentiles across groups using modified z-scores (Mean Absolute Deviation - MAD).
}
