% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvFUN.r
\name{rvFUN}
\alias{rvFUN}
\title{A daily ranked variance function}
\usage{
rvFUN(x, obs, date, group, ell = 5, reflective = TRUE, by.day = TRUE,
  theta = NA, tau = NA)
}
\arguments{
\item{obs}{This is the numeric column under investigation.}

\item{date}{This is a required column with the format 'YYYY-MM-DD HH:MM:SS'.}

\item{group}{This is the first grouping value of the data, with the default being 'site'.}

\item{reflective}{This decides whether to use all selected data (TRUE), or to use data from the latest day (live). Defaults to TRUE.}

\item{by.day}{This is an option for where analysis uses data from only that day ('cross', default), or uses all selected data ('auto', FALSE). If selecting auto, the last five days will be used.}

\item{theta}{This sets the test threshold on whether to flag the data. Defaults to NA (no flags given).}

\item{tau}{This sets the day threshold on whether to flag the data given consistent theta flags. Defaults to NA (no flags given).}
}
\description{
This function gives the daily ranked variance for an observation, relative to a pool of other observations.
}
\examples{
rvFUN()
}
