% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementaryFUN.r
\name{vlFUN}
\alias{vlFUN}
\title{A test to see how long it has been since visits.}
\usage{
vlFUN(x, date, last.visit = "last.visit", theta = 60)
}
\arguments{
\item{date}{This is the date column (in YYYY-MM-DD).}

\item{last.visit}{This is the column of the date of last visit to the site (in YYYY-MM-DD).}

\item{theta}{This is the threshold of when length between site visits is 'high' (in days).}
}
\description{
A test to see how long it has been since visits.
}
\examples{
vlFUN()
}
