% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cross.r
\name{crossFUN}
\alias{crossFUN}
\title{A function to set up data ready to perform "cross" analysis.}
\usage{
crossFUN(x, date, obs, group)
}
\arguments{
\item{obs}{This is the column under observation.}

\item{group}{This is the id of each set of data being compared.}
}
\description{
This function reformats the data relative to cross analysis, where time is held steady.
}
\examples{
crossFUN()
}
