% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{distFUN}
\alias{distFUN}
\title{A selection of functions used within other functions in the package}
\usage{
distFUN(x, date)
}
\arguments{
\item{date}{Date column}
}
\description{
A function for creating dissimilarity plots using Spearman rank correlations.
}
\examples{
distFUN()
}
