% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{distFUN}
\alias{distFUN}
\title{A selection of functions used within other functions in the package}
\usage{
distFUN(x, omit.cols = NA)
}
\arguments{
\item{omit.cols}{This is a list of the columns to remove from analysis if needed. Default is}
}
\description{
A function for creating dissimilarity plots using Spearman rank correlations.
}
\examples{
distFUN()
}
