% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementaryFUN.r
\name{frFUN}
\alias{frFUN}
\title{A test to check where the flow rate is too low}
\usage{
frFUN(x, theta = 1.5)
}
\arguments{
\item{theta}{This is the threshold of what 'low' flow is.}
}
\description{
A test to check where the flow rate is too low
}
\examples{
frFUN()
}
