% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementaryFUN.r
\name{hcFUN}
\alias{hcFUN}
\title{A test to check if values are too high.}
\usage{
hcFUN(x, obs = "value", theta = 50)
}
\arguments{
\item{obs}{This is the column of interest, with default being 'value'.}

\item{theta}{This is the threshold of what a 'high' value is.}
}
\description{
This function gives binary output on whether values are over a defined threshold.
}
\examples{
hcFUN()
}
