% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{ksplotFUN}
\alias{ksplotFUN}
\title{A daily ranked variance function}
\usage{
ksplotFUN(x)
}
\arguments{
\item{reflective}{This decides whether to use a selected time period of data (reflective), or to use data from the latest day (live). Defaults to TRUE.}

\item{plot}{This decides whether to plot the data. Defaults to TRUE.}
}
\description{
This function gives the daily ranked variance for an observation, relative to a pool of other observations.
}
\examples{
ksplotFUN()
}

