% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{networkmedianFUN}
\alias{networkmedianFUN}
\title{This function gives the proxy type of network median.}
\usage{
networkmedianFUN(x, group, obs)
}
\arguments{
\item{group}{This is the identifying variable.}

\item{obs}{This is the column being examined.}
}
\description{
This function gives the proxy type of network median.
}
\examples{
networkmedianFUN()
}
