% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_within_functions.r
\name{rollingKStest}
\alias{rollingKStest}
\title{This function gives the rolling KS function}
\usage{
rollingKStest(z, obs, proxy, window = 1440)
}
\arguments{
\item{obs}{This is the column being examined.}

\item{proxy}{This is the comparison variable}

\item{window.}{This defines the size of the window to sample data from}
}
\description{
This function gives the rolling KS function
}
\examples{
rollingKStest()
}
