% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvFUN.r
\name{rvFUN}
\alias{rvFUN}
\title{A daily ranked variance function}
\usage{
rvFUN(x, obs, reflective = TRUE, plot = TRUE, date.start = "2016-07-01",
  date.end = "2016-07-10", theta = NA, tau = NA)
}
\arguments{
\item{reflective}{This decides whether to use a selected time period of data (reflective), or to use data from the latest day (live). Defaults to TRUE.}

\item{plot}{This decides whether to plot the data. Defaults to TRUE.}

\item{date.start}{This sets the start of the selected data, if reflective is true. Defaults to '2016-07-01'.}

\item{date.end}{This sets the end of the selected data, if reflective is true. Defaults to '2016-07-10'}

\item{theta}{This sets the test threshold on whether to flag the data. Defaults to NA (no flags given).}

\item{tau}{This sets the day threshold on whether to flag the data given consistent theta flags. Defaults to NA (no flags given).}
}
\description{
This function gives the daily ranked variance for an observation, relative to a pool of other observations.
}
\examples{
rvFUN()
}

