% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementaryFUN.r
\name{vlFUN}
\alias{vlFUN}
\title{A test to see how long it has been since visits.}
\usage{
vlFUN(x, theta = 60)
}
\arguments{
\item{theta}{This is the threshold of when length between site visits is 'high' (in days).}
}
\description{
A test to see how long it has been since visits.
}
\examples{
vlFUN()
}
