% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementaryFUN.r
\name{vlFUN}
\alias{vlFUN}
\title{A daily ranked variance function}
\usage{
vlFUN(x, length.threshold = 60)
}
\arguments{
\item{reflective}{This decides whether to use a selected time period of data (reflective), or to use data from the latest day (live). Defaults to TRUE.}

\item{plot}{This decides whether to plot the data. Defaults to TRUE.}
}
\description{
This function gives the daily ranked variance for an observation, relative to a pool of other observations.
}
\examples{
vlFUN()
}

